+++ app/trakt_tools.py
@@ -2,6 +2,7 @@ import config_tools
 from urllib.parse import urlparse
 import plex_tools
 import trakt
+import os
 
 def trakt_get_movies(config_path, plex, data, is_userlist=True):
     config_tools.TraktClient(config_path)
@@ -18,14 +19,21 @@ def trakt_get_movies(config_path, plex, data, is_userlist=True):
     title_ids = [m.pk[1] for m in trakt_list_items if isinstance(m, trakt.objects.movie.Movie)]
 
     imdb_map = {}
+    plex_tools.create_cache(config_path)
     if title_ids:
         for item in plex.Library.all():
-            guid = urlparse(item.guid)
-            item_type = guid.scheme.split('.')[-1]
-            if item_type == 'imdb':
-                imdb_id = guid.netloc
+            item_type = urlparse(item.guid).scheme.split('.')[-1]
+            if item_type == 'plex':
+                # Check cache for imdb_id
+                imdb_id = plex_tools.query_cache(config_path, item.guid, 'imdb_id')
+                if not imdb_id:
+                    imdb_id, tmdb_id = plex_tools.alt_id_lookup(plex, item)
+                    print("| Cache | + | {} | {} | {} | {}".format(item.guid, imdb_id, tmdb_id, item.title))
+                    plex_tools.update_cache(config_path, item.guid, imdb_id=imdb_id, tmdb_id=tmdb_id)
+            elif item_type == 'imdb':
+                imdb_id = urlparse(item.guid).netloc
             elif item_type == 'themoviedb':
-                tmdb_id = guid.netloc
+                tmdb_id = urlparse(item.guid).netloc
                 # lookup can sometimes return a list
                 lookup = trakt.Trakt['search'].lookup(tmdb_id, 'tmdb', 'movie')
                 if lookup: